Rule request: UseClosableForHttpClient #97 Added remoting as category, moved applicable rules from common to remoting
