+++ src/main/resources/category/java/categories.properties
@@ -1,6 +1,7 @@
 rulesets.filenames=\
     category/java/common.xml,\
     category/java/common-std.xml,\
+    category/java/remoting.xml,\
     category/java/concurrent.xml,\
     category/java/spring.xml,\
     category/java/sql.xml
