+++ src/main/resources/category/java/spring.xml
@@ -196,7 +196,7 @@ count(.//PrimaryExpression/PrimaryPrefix/Name[ends-with(@Image,'.clear')])=0]
     </rule>
 
     <rule name="AvoidExpressionsInCacheable" class="net.sourceforge.pmd.lang.rule.XPathRule" deprecated="false" dfa="false" language="java" message="Avoid SpEL-expression for computing Cacheable key" typeResolution="true"
-          externalInfoUrl="${doc_root}/JavaCodePerformance.md#tmsu12">
+          externalInfoUrl="${doc_root}/JavaCodePerformance.md#improper-caching">
         <description>Spring Expression Language (SpEL) expression is used for computing the key dynamically. Problem: evaluating the expression language is expensive, on every call.&#13;
             Solution: use a custom KeyGenerator: keyGenerator=... instead of key=...
             (jpinpoint-rules)</description>
@@ -222,7 +222,7 @@ class Bad1 {
     </rule>
 
     <rule name="SynchronizeForKeyInCacheable" class="net.sourceforge.pmd.lang.rule.XPathRule" deprecated="false" dfa="false" language="java" message="Synchronize access for each key in @Cacheable" typeResolution="true"
-          externalInfoUrl="${doc_root}/JavaCodePerformance.md#tmsu12">
+          externalInfoUrl="${doc_root}/JavaCodePerformance.md#improper-caching">
         <description>The cache by default allows multiple threads accessing by the same key. Problem: if the value of the key is not available from the cache, it will be fetched/computed by multiple threads while only one time is needed.&#13;
             Solution: Let only the first accessing thread fetch/compute the value and block others until the value is in the cache. Add attribute sync = "true" to achieve this. (Assuming the cache implementation supports it.)
             (jpinpoint-rules)</description>
