+++ src/test/java/com/jpinpoint/perf/lang/java/ruleset/remoting/JAXBContextCreatedForEachMethodCallTest.java
@@ -1,4 +1,4 @@
-package com.jpinpoint.perf.lang.java.ruleset.common;
+package com.jpinpoint.perf.lang.java.ruleset.remoting;
 
 import net.sourceforge.pmd.testframework.PmdRuleTst;
