+++ src/test/java/com/jpinpoint/perf/lang/java/ruleset/remoting/UseCloseableForHttpClientTest.java
@@ -1,0 +1,6 @@
+package com.jpinpoint.perf.lang.java.ruleset.remoting;
+
+import net.sourceforge.pmd.testframework.PmdRuleTst;
+
+public class UseCloseableForHttpClientTest extends PmdRuleTst {
+}
