+++ src/test/resources/com/jpinpoint/perf/lang/java/ruleset/remoting/xml/AvoidDeprecatedHttpConnectors.xml
